fix: revise ReplyModal
