+++ src/assets/styles/_follow.scss
@@ -5,7 +5,7 @@
   border-right: 1px solid var(--border);
 
   &__no__content {
-    margin-top: 30%;
+    margin-top: 25%;
   }
 
   &__header {
