+++ src/assets/styles/_tweet.scss
@@ -90,7 +90,7 @@
     border-top: 10px solid var(--border);
 
     &__blank {
-      margin: 24px auto;
+      margin-top: 25%;
       text-align: center;
     }
