+++ src/assets/styles/_user-like.scss
@@ -2,7 +2,7 @@
   height: 47vh;
 
   &__no__content {
-    margin-top: 30%;
+    margin-top: 25%;
   }
 
   &__tweet {
