+++ src/assets/styles/_user-tweet.scss
@@ -6,7 +6,7 @@
     height: 45vh;
 
     &__blank {
-      margin: 24px auto;
+      margin-top: 25%;
       text-align: center;
     }
