+++ src/components/UserReply.vue
@@ -138,20 +138,15 @@ export default {
           icon: 'warning',
           title: '刪除無法復原，確認刪除？',
           showCancelButton: true,
-          cancelButtonColor: '#fc5a5a',
+          cancelButtonColor: '#50b5ff',
           cancelButtonText: '取消',
-          confirmButtonColor: '#50b5ff',
+          confirmButtonColor: '#fc5a5a',
           confirmButtonText: '確認'
         })
 
-        if (result.isConfirmed) {
-          Toast.fire({
-            icon: 'success',
-            title: '成功刪除回覆'
-          })
-        } else {
+        if (!result.isConfirmed) {
           this.isProcessing = false
-          return false
+          return
         }
 
         const { data } = await tweetsAPI.deleteReply({ tweet_id, id })
@@ -163,6 +158,12 @@ export default {
         this.repliedTweets = this.repliedTweets.filter(
           (reply) => reply.id !== id
         )
+
+        Toast.fire({
+          icon: 'success',
+          title: '成功刪除回覆'
+        })
+
         this.isProcessing = false
       } catch (error) {
         this.isProcessing = false
