+++ src/views/Profile.vue
@@ -48,12 +48,15 @@ export default {
   },
   created() {
     const { id } = this.$route.params
+
     this.fetchUsers(id)
     this.fetchFollowers(id)
   },
   beforeRouteUpdate(to, from, next) {
     const { id } = to.params
+
     this.fetchUsers(id)
+    this.fetchFollowers(id)
     next()
   },
   methods: {
@@ -73,7 +76,9 @@ export default {
         this.isLoading = false
       } catch (error) {
         this.isLoading = false
+
         console.log(error)
+
         Toast.fire({
           icon: 'error',
           title: '無法取得使用者資料'
@@ -93,7 +98,9 @@ export default {
         this.isLoading = false
       } catch (error) {
         this.isLoading = false
+
         console.log(error)
+
         Toast.fire({
           icon: 'error',
           title: '無法取得跟隨您的使用者資料'
@@ -108,6 +115,7 @@ export default {
     renderProfile: function () {
       if (this.currentUser.id !== this.user.id) {
         const { id } = this.$route.params
+
         this.fetchFollowers(id)
         this.fetchUsers(id)
       }
